Added support for CSV files
Now can use CSV files for setting the color. The --col options can
be either numbers or column names. Internally, pandas DataFrame
is used for reading the CSV files and added to the dependencies.
There is no change to the existing behaviour.

In addition, this allows external data to be blended into the JSON
files generated for chemischope.